Create the layout of strategy table.
