+++ src/components/StrategyPage/StrategyContainer.tsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import StrategyInformation from './StrategyInformation';
 import { StrategyTypes } from '../../enums/strategies';
-import StrategyTable from './StrategyTable';
+import StrategyTable from './StrategyTable/StrategyTable';
 import { StrategyWrapper } from './styled';
 import { Col, Row } from 'antd';
 import { StandardText } from '../../reducers/client';
@@ -26,7 +26,7 @@ class StrategyContainer extends PureComponent<StrategyContainerProps> {
             <StrategyInformation {...information} type={type} />
           </Col>
           <Col span={12}>
-            <StrategyTable strategies={strategies} />
+            <StrategyTable strategies={strategies} type={type} />
           </Col>
         </Row>
       </StrategyWrapper>
